+++ nextjs/src/components/cards/card-slider-explanation.tsx
@@ -14,7 +14,10 @@ const CardSliderExplanation: React.FC<CardExplanationProps> = ({
   children,
 }) => {
   return (
-    <div className="snap-center hidden sm:flex flex-col items-start gap-8 pt-6 self-stretch px-6 pb-8 rounded-xl bg-white shadow-2 max-w-md">
+    <div
+      data-component="CardSliderExplanation"
+      className="snap-center hidden sm:flex flex-col items-start gap-8 pt-6 self-stretch px-6 pb-8 rounded-xl bg-white shadow-2 max-w-md"
+    >
       <h3 className="text-30 font-semibold leading-none text-biscay">
         {title}
       </h3>