+++ nextjs/src/components/nav-bar/nav-mobile.tsx
@@ -1,4 +1,4 @@
-import React, { useState, type LegacyRef } from "react"
+import React, { useState, type Ref } from "react"
 import clsx from "clsx"
 import { Transition } from "@headlessui/react"
 import { useClickAway } from "@uidotdev/usehooks"
@@ -6,10 +6,11 @@ import { useClickAway } from "@uidotdev/usehooks"
 import IconHamburgerMenu from "../icons/icon-hamburger-menu"
 import Link from "next/link"
 import ButtonLink from "../link-button"
-import IconChevronRight from "../icons/icon-chevron-right"
 import Logo from "../logo"
 import TopbarActions from "../topbar-actions"
 import type { NavItem } from "./nav"
+import NavMobileItem from "./nav-mobile-item"
+import IconChevronLeft from "../icons/icon-chevron-left"
 
 type NavMobileProps = {
   navItems: NavItem[]
@@ -17,26 +18,37 @@ type NavMobileProps = {
 
 const NavMobile: React.FC<NavMobileProps> = ({ navItems }) => {
   const [isOpen, setIsOpen] = useState(false)
+  const [expand, setExpand] = useState<number | undefined>(undefined)
   const ref = useClickAway(() => {
     setIsOpen(false)
-  }) as LegacyRef<HTMLButtonElement> | undefined
+    setExpand(undefined)
+  }) as Ref<HTMLDivElement> | undefined
 
   return (
-    <div className="lg:hidden relative">
+    <div className="lg:hidden relative" ref={ref}>
       <nav className="grid grid-cols-3 justify-between h-20 px-4 ">
-        <button onClick={() => setIsOpen(!isOpen)} ref={ref}>
-          <IconHamburgerMenu />
-        </button>
+        {expand === undefined ? (
+          <button onClick={() => setIsOpen(!isOpen)}>
+            <IconHamburgerMenu />
+          </button>
+        ) : (
+          <button onClick={() => setExpand(undefined)}>
+            <IconChevronLeft className="text-white w-6 h-6" />
+          </button>
+        )}
         <Link href="/theme" className="h-20 flex justify-center">
           <Logo color="biscay" variant="icon" />
         </Link>
         <TopbarActions />
       </nav>
+
       <Transition show={isOpen}>
         <ul
           className={clsx([
             // Base styles
-            "absolute px-4 pt-4 pb-6 left-0 right-0 top-20 -translate-y-1 w-full transition ease-in-out grid bg-sapphire",
+            "absolute px-4 pt-4 pb-6 left-0 right-0 top-20  w-full grid bg-sapphire",
+            // Shared transition styles
+            "transition ease-out -translate-y-1",
             // Shared closed styles
             "data-[closed]:opacity-0",
             // Entering styles
@@ -45,13 +57,23 @@ const NavMobile: React.FC<NavMobileProps> = ({ navItems }) => {
             "data-[leave]:duration-300 data-[leave]:data-[closed]:-translate-y-4",
           ])}
         >
-          <li>
-            <Link className="flex justify-between py-4" href="/theme">
-              Adfinis.com
-              <IconChevronRight />
-            </Link>
-          </li>
-          <li>
+          {navItems.map((item, index) => (
+            <li
+              onClick={() => setExpand(index)}
+              key={index}
+              className={clsx([
+                { hidden: expand !== index && expand !== undefined },
+              ])}
+            >
+              <NavMobileItem
+                key={index}
+                item={item}
+                expand={expand === index}
+              />
+            </li>
+          ))}
+
+          <li className="mt-8">
             <ButtonLink href={"/theme"} variant={"cta"}>
               Get started!
             </ButtonLink>